FIx INT type

+++ project_1/project_1.sql
@@ -1,38 +1,38 @@
 -- create initial tables based on 3NF
 
 CREATE TABLE Agency (
-    agency_code TINYINT PRIMARY KEY,
+    agency_code SMALLINT PRIMARY KEY,
     agency_name VARCHAR(100) NOT NULL,
     agency_abbr VARCHAR(20) NOT NULL
 );
 
 CREATE TABLE LoanType (
-    loan_type TINYINT PRIMARY KEY,
+    loan_type SMALLINT PRIMARY KEY,
     loan_type_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE PropertyType (
-    property_type TINYINT PRIMARY KEY,
+    property_type SMALLINT PRIMARY KEY,
     property_type_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE LoanPurpose (
-    loan_purpose TINYINT PRIMARY KEY,
+    loan_purpose SMALLINT PRIMARY KEY,
     loan_purpose_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE OwnerOccupancy (
-    owner_occupancy TINYINT PRIMARY KEY,
+    owner_occupancy SMALLINT PRIMARY KEY,
     owner_occupancy_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE Preapproval (
-    preapproval TINYINT PRIMARY KEY,
+    preapproval SMALLINT PRIMARY KEY,
     preapproval_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE ActionTaken (
-    action_taken TINYINT PRIMARY KEY,
+    action_taken SMALLINT PRIMARY KEY,
     action_taken_name VARCHAR(100) NOT NULL
 );
 
@@ -56,19 +56,19 @@ CREATE TABLE County (
 );
 
 CREATE TABLE Ethnicity (
-    ethnicity_code TINYINT PRIMARY KEY,
+    ethnicity_code SMALLINT PRIMARY KEY,
     ethnicity_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE Race (
-    race_code TINYINT PRIMARY KEY,
+    race_code SMALLINT PRIMARY KEY,
     race_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE ApplicantRace (
     ID INTEGER,
-    race_number TINYINT,
-    race_code TINYINT NOT NULL,
+    race_number SMALLINT,
+    race_code SMALLINT NOT NULL,
     PRIMARY KEY (ID, race_number),
     FOREIGN KEY (ID) REFERENCES LoanApplication(ID),
     FOREIGN KEY (race_code) REFERENCES Race(race_code)
@@ -76,49 +76,49 @@ CREATE TABLE ApplicantRace (
 
 CREATE TABLE CoApplicantRace (
     ID INTEGER,
-    race_number TINYINT,
-    race_code TINYINT NOT NULL,
+    race_number SMALLINT,
+    race_code SMALLINT NOT NULL,
     PRIMARY KEY (ID, race_number),
     FOREIGN KEY (ID) REFERENCES LoanApplication(ID),
     FOREIGN KEY (race_code) REFERENCES Race(race_code)
 );
 
 CREATE TABLE Sex (
-    sex_code TINYINT PRIMARY KEY,
+    sex_code SMALLINT PRIMARY KEY,
     sex_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE PurchaserType (
-    purchaser_type TINYINT PRIMARY KEY,
+    purchaser_type SMALLINT PRIMARY KEY,
     purchaser_type_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE DenialReason (
-    denial_reason_code TINYINT PRIMARY KEY,
+    denial_reason_code SMALLINT PRIMARY KEY,
     denial_reason_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE DenialReasons (
     ID INTEGER,
-    reason_number TINYINT,
-    denial_reason_code TINYINT NOT NULL,
+    reason_number SMALLINT,
+    denial_reason_code SMALLINT NOT NULL,
     PRIMARY KEY (ID, reason_number),
     FOREIGN KEY (ID) REFERENCES LoanApplication(ID),
     FOREIGN KEY (denial_reason_code) REFERENCES DenialReason(denial_reason_code)
 );
 
 CREATE TABLE HOEPAStatus (
-    hoepa_status TINYINT PRIMARY KEY,
+    hoepa_status SMALLINT PRIMARY KEY,
     hoepa_status_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE LienStatus (
-    lien_status TINYINT PRIMARY KEY,
+    lien_status SMALLINT PRIMARY KEY,
     lien_status_name VARCHAR(100) NOT NULL
 );
 
 CREATE TABLE EditStatus (
-    edit_status TINYINT PRIMARY KEY,
+    edit_status SMALLINT PRIMARY KEY,
     edit_status_name VARCHAR(100) NOT NULL
 );
 
@@ -143,7 +143,7 @@ CREATE TABLE Location (
 CREATE TABLE RespondentAgency (
     as_of_year INTEGER,
     respondent_id VARCHAR(10),
-    agency_code TINYINT,
+    agency_code SMALLINT,
     PRIMARY KEY (as_of_year, respondent_id),
     FOREIGN KEY (agency_code) REFERENCES Agency(agency_code)
 );
@@ -152,26 +152,26 @@ CREATE TABLE LoanApplication (
     ID INTEGER PRIMARY KEY,
     as_of_year INTEGER NOT NULL,
     respondent_id VARCHAR(10) NOT NULL,
-    loan_type TINYINT NOT NULL,
-    property_type TINYINT NOT NULL,
-    loan_purpose TINYINT NOT NULL,
-    owner_occupancy TINYINT NOT NULL,
+    loan_type SMALLINT NOT NULL,
+    property_type SMALLINT NOT NULL,
+    loan_purpose SMALLINT NOT NULL,
+    owner_occupancy SMALLINT NOT NULL,
     loan_amount_000s NUMERIC,
-    preapproval TINYINT NOT NULL,
-    action_taken TINYINT NOT NULL,
+    preapproval SMALLINT NOT NULL,
+    action_taken SMALLINT NOT NULL,
     location_id INTEGER NOT NULL,
-    applicant_ethnicity TINYINT NOT NULL,
-    co_applicant_ethnicity TINYINT,
-    applicant_sex TINYINT NOT NULL,
-    co_applicant_sex TINYINT,
+    applicant_ethnicity SMALLINT NOT NULL,
+    co_applicant_ethnicity SMALLINT,
+    applicant_sex SMALLINT NOT NULL,
+    co_applicant_sex SMALLINT,
     applicant_income_000s NUMERIC,
-    purchaser_type TINYINT,
+    purchaser_type SMALLINT,
     rate_spread VARCHAR(10),
-    hoepa_status TINYINT,
-    lien_status TINYINT,
-    edit_status TINYINT,
+    hoepa_status SMALLINT,
+    lien_status SMALLINT,
+    edit_status SMALLINT,
     sequence_number VARCHAR(20),
-    application_date_indicator TINYINT,
+    application_date_indicator SMALLINT,
     FOREIGN KEY (as_of_year, respondent_id) REFERENCES RespondentAgency(as_of_year, respondent_id),
     FOREIGN KEY (loan_type) REFERENCES LoanType(loan_type),
     FOREIGN KEY (property_type) REFERENCES PropertyType(property_type),
@@ -193,7 +193,7 @@ CREATE TABLE LoanApplication (
 -- fill lookup tables with distinct values
 INSERT INTO Agency (agency_code, agency_name, agency_abbr)
 SELECT DISTINCT
-    NULLIF(agency_code, '')::TINYINT,
+    NULLIF(agency_code, '')::SMALLINT,
     NULLIF(agency_name, ''),
     NULLIF(agency_abbr, '')
 FROM Preliminary
@@ -201,42 +201,42 @@ WHERE agency_code != '';
 
 INSERT INTO LoanType (loan_type, loan_type_name)
 SELECT DISTINCT
-    NULLIF(loan_type, '')::TINYINT,
+    NULLIF(loan_type, '')::SMALLINT,
     NULLIF(loan_type_name, '')
 FROM Preliminary
 WHERE loan_type != '';
 
 INSERT INTO PropertyType (property_type, property_type_name)
 SELECT DISTINCT
-    NULLIF(property_type, '')::TINYINT,
+    NULLIF(property_type, '')::SMALLINT,
     NULLIF(property_type_name, '')
 FROM Preliminary
 WHERE property_type != '';
 
 INSERT INTO LoanPurpose (loan_purpose, loan_purpose_name)
 SELECT DISTINCT
-    NULLIF(loan_purpose, '')::TINYINT,
+    NULLIF(loan_purpose, '')::SMALLINT,
     NULLIF(loan_purpose_name, '')
 FROM Preliminary
 WHERE loan_purpose != '';
 
 INSERT INTO OwnerOccupancy (owner_occupancy, owner_occupancy_name)
 SELECT DISTINCT
-    NULLIF(owner_occupancy, '')::TINYINT,
+    NULLIF(owner_occupancy, '')::SMALLINT,
     NULLIF(owner_occupancy_name, '')
 FROM Preliminary
 WHERE owner_occupancy != '';
 
 INSERT INTO Preapproval (preapproval, preapproval_name)
 SELECT DISTINCT
-    NULLIF(preapproval, '')::TINYINT,
+    NULLIF(preapproval, '')::SMALLINT,
     NULLIF(preapproval_name, '')
 FROM Preliminary
 WHERE preapproval != '';
 
 INSERT INTO ActionTaken (action_taken, action_taken_name)
 SELECT DISTINCT
-    NULLIF(action_taken, '')::TINYINT,
+    NULLIF(action_taken, '')::SMALLINT,
     NULLIF(action_taken_name, '')
 FROM Preliminary
 WHERE action_taken != '';
@@ -266,134 +266,134 @@ WHERE county_code != '' AND state_code != '';
 
 INSRT INTO Ethnicity (ethnicity_code, ethnicity_name)
 SELECT DISTINCT
-    NULLIF(applicant_ethnicity, '')::TINYINT,
+    NULLIF(applicant_ethnicity, '')::SMALLINT,
     NULLIF(applicant_ethnicity_name, '')
 FROM Preliminary
 WHERE applicant_ethnicity != ''
 UNION
 SELECT DISTINCT
-    NULLIF(co_applicant_ethnicity, '')::TINYINT,
+    NULLIF(co_applicant_ethnicity, '')::SMALLINT,
     NULL(co_applicant_ethnicity_name, '')
 FROM Preliminary
 WHERE co_applicant_ethnicity != '';
 
 INSERT INTO Race (race_code, race_name)
 SELECT DISTINCT 
-    NULLIF(applicant_race_1, '')::TINYINT, 
+    NULLIF(applicant_race_1, '')::SMALLINT, 
     NULLIF(applicant_race_name_1, '')
 FROM Preliminary
 WHERE applicant_race_1 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(applicant_race_2, '')::TINYINT, 
+    NULLIF(applicant_race_2, '')::SMALLINT, 
     NULLIF(applicant_race_name_2, '')
 FROM Preliminary
 WHERE applicant_race_2 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(applicant_race_3, '')::TINYINT, 
+    NULLIF(applicant_race_3, '')::SMALLINT, 
     NULLIF(applicant_race_name_3, '')
 FROM Preliminary
 WHERE applicant_race_3 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(applicant_race_4, '')::TINYINT, 
+    NULLIF(applicant_race_4, '')::SMALLINT, 
     NULLIF(applicant_race_name_4, '')
 FROM Preliminary
 WHERE applicant_race_4 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(applicant_race_5, '')::TINYINT, 
+    NULLIF(applicant_race_5, '')::SMALLINT, 
     NULLIF(applicant_race_name_5, '')
 FROM Preliminary
 WHERE applicant_race_5 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_race_1, '')::TINYINT, 
+    NULLIF(co_applicant_race_1, '')::SMALLINT, 
     NULLIF(co_applicant_race_name_1, '')
 FROM Preliminary
 WHERE co_applicant_race_1 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_race_2, '')::TINYINT, 
+    NULLIF(co_applicant_race_2, '')::SMALLINT, 
     NULLIF(co_applicant_race_name_2, '')
 FROM Preliminary
 WHERE co_applicant_race_2 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_race_3, '')::TINYINT, 
+    NULLIF(co_applicant_race_3, '')::SMALLINT, 
     NULLIF(co_applicant_race_name_3, '')
 FROM Preliminary
 WHERE co_applicant_race_3 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_race_4, '')::TINYINT, 
+    NULLIF(co_applicant_race_4, '')::SMALLINT, 
     NULLIF(co_applicant_race_name_4, '')
 FROM Preliminary
 WHERE co_applicant_race_4 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_race_5, '')::TINYINT, 
+    NULLIF(co_applicant_race_5, '')::SMALLINT, 
     NULLIF(co_applicant_race_name_5, '')
 FROM Preliminary
 WHERE co_applicant_race_5 != '';
 
 INSERT INTO Sex (sex_code, sex_name)
 SELECT DISTINCT 
-    NULLIF(applicant_sex, '')::TINYINT, 
+    NULLIF(applicant_sex, '')::SMALLINT, 
     NULLIF(applicant_sex_name, '')
 FROM Preliminary
 WHERE applicant_sex != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(co_applicant_sex, '')::TINYINT, 
+    NULLIF(co_applicant_sex, '')::SMALLINT, 
     NULLIF(co_applicant_sex_name, '')
 FROM Preliminary
 WHERE co_applicant_sex != '';
 
 INSERT INTO PurchaserType (purchaser_type, purchaser_type_name)
 SELECT DISTINCT 
-    NULLIF(purchaser_type, '')::TINYINT, 
+    NULLIF(purchaser_type, '')::SMALLINT, 
     NULLIF(purchaser_type_name, '')
 FROM Preliminary
 WHERE purchaser_type != '';
 
 INSERT INTO DenialReason (denial_reason_code, denial_reason_name)
 SELECT DISTINCT 
-    NULLIF(denial_reason_1, '')::TINYINT, 
+    NULLIF(denial_reason_1, '')::SMALLINT, 
     NULLIF(denial_reason_name_1, '')
 FROM Preliminary
 WHERE denial_reason_1 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(denial_reason_2, '')::TINYINT, 
+    NULLIF(denial_reason_2, '')::SMALLINT, 
     NULLIF(denial_reason_name_2, '')
 FROM Preliminary
 WHERE denial_reason_2 != ''
 UNION
 SELECT DISTINCT 
-    NULLIF(denial_reason_3, '')::TINYINT, 
+    NULLIF(denial_reason_3, '')::SMALLINT, 
     NULLIF(denial_reason_name_3, '')
 FROM Preliminary
 WHERE denial_reason_3 != '';
 
 INSERT INTO HOEPAStatus (hoepa_status, hoepa_status_name)
 SELECT DISTINCT 
-    NULLIF(hoepa_status, '')::TINYINT, 
+    NULLIF(hoepa_status, '')::SMALLINT, 
     NULLIF(hoepa_status_name, '')
 FROM Preliminary
 WHERE hoepa_status != '';
 
 INSERT INTO LienStatus (lien_status, lien_status_name)
 SELECT DISTINCT 
-    NULLIF(lien_status, '')::TINYINT, 
+    NULLIF(lien_status, '')::SMALLINT, 
     NULLIF(lien_status_name, '')
 FROM Preliminary
 WHERE lien_status != '';
 
 INSERT INTO EditStatus (edit_status, edit_status_name)
 SELECT DISTINCT 
-    NULLIF(edit_status, '')::TINYINT, 
+    NULLIF(edit_status, '')::SMALLINT, 
     NULLIF(edit_status_name, '')
 FROM Preliminary
 WHERE edit_status != '';
@@ -402,75 +402,75 @@ INSERT INTO RespondentAgency (as_of_year, respondent_id, agency_code)
 SELECT DISTINCT 
     NULLIF(as_of_year, '')::INTEGER, 
     NULLIF(respondent_id, ''), 
-    NULLIF(agency_code, '')::TINYINT
+    NULLIF(agency_code, '')::SMALLINT
 FROM Preliminary
 WHERE as_of_year != '' AND respondent_id != '' AND agency_code != '';
 
 -- fill ApplicantRace table
 INSERT INTO ApplicantRace (ID, race_number, race_code)
-SELECT ID, 1, NULLIF(applicant_race_1, '')::TINYINT
+SELECT ID, 1, NULLIF(applicant_race_1, '')::SMALLINT
 FROM Preliminary
 WHERE applicant_race_1 != '';
 
 INSERT INTO ApplicantRace (ID, race_number, race_code)
-SELECT ID, 2, NULLIF(applicant_race_2, '')::TINYINT
+SELECT ID, 2, NULLIF(applicant_race_2, '')::SMALLINT
 FROM Preliminary
 WHERE applicant_race_2 != '';
 
 INSERT INTO ApplicantRace (ID, race_number, race_code)
-SELECT ID, 3, NULLIF(applicant_race_3, '')::TINYINT
+SELECT ID, 3, NULLIF(applicant_race_3, '')::SMALLINT
 FROM Preliminary
 WHERE applicant_race_3 != '';
 
 INSERT INTO ApplicantRace (ID, race_number, race_code)
-SELECT ID, 4, NULLIF(applicant_race_4, '')::TINYINT
+SELECT ID, 4, NULLIF(applicant_race_4, '')::SMALLINT
 FROM Preliminary
 WHERE applicant_race_4 != '';
 
 INSERT INTO ApplicantRace (ID, race_number, race_code)
-SELECT ID, 5, NULLIF(applicant_race_5, '')::TINYINT
+SELECT ID, 5, NULLIF(applicant_race_5, '')::SMALLINT
 FROM Preliminary
 WHERE applicant_race_5 != '';
 
 -- fill CoApplicantRace table
 INSERT INTO CoApplicantRace (ID, race_number, race_code)
-SELECT ID, 1, NULLIF(co_applicant_race_1, '')::TINYINT
+SELECT ID, 1, NULLIF(co_applicant_race_1, '')::SMALLINT
 FROM Preliminary
 WHERE co_applicant_race_1 != '';
 
 INSERT INTO CoApplicantRace (ID, race_number, race_code)
-SELECT ID, 2, NULLIF(co_applicant_race_2, '')::TINYINT
+SELECT ID, 2, NULLIF(co_applicant_race_2, '')::SMALLINT
 FROM Preliminary
 WHERE co_applicant_race_2 != '';
 
 INSERT INTO CoApplicantRace (ID, race_number, race_code)
-SELECT ID, 3, NULLIF(co_applicant_race_3, '')::TINYINT
+SELECT ID, 3, NULLIF(co_applicant_race_3, '')::SMALLINT
 FROM Preliminary
 WHERE co_applicant_race_3 != '';
 
 INSERT INTO CoApplicantRace (ID, race_number, race_code)
-SELECT ID, 4, NULLIF(co_applicant_race_4, '')::TINYINT
+SELECT ID, 4, NULLIF(co_applicant_race_4, '')::SMALLINT
 FROM Preliminary
 WHERE co_applicant_race_4 != '';
 
 INSERT INTO CoApplicantRace (ID, race_number, race_code)
-SELECT ID, 5, NULLIF(co_applicant_race_5, '')::TINYINT
+SELECT ID, 5, NULLIF(co_applicant_race_5, '')::SMALLINT
 FROM Preliminary
 WHERE co_applicant_race_5 != '';
 
 -- fill DenialReasons table
 INSERT INTO DenialReasons (ID, reason_number, denial_reason_code)
-SELECT ID, 1, NULLIF(denial_reason_1, '')::TINYINT
+SELECT ID, 1, NULLIF(denial_reason_1, '')::SMALLINT
 FROM Preliminary
 WHERE denial_reason_1 != '';
 
 INSERT INTO DenialReasons (ID, reason_number, denial_reason_code)
-SELECT ID, 2, NULLIF(denial_reason_2, '')::TINYINT
+SELECT ID, 2, NULLIF(denial_reason_2, '')::SMALLINT
 FROM Preliminary
 WHERE denial_reason_2 != '';
 
 INSERT INTO DenialReasons (ID, reason_number, denial_reason_code)
-SELECT ID, 3, NULLIF(denial_reason_3, '')::TINYINT
+SELECT ID, 3, NULLIF(denial_reason_3, '')::SMALLINT
 FROM Preliminary
 WHERE denial_reason_3 != '';
 
@@ -558,26 +558,26 @@ SELECT
     p.ID,
     NULLIF(p.as_of_year, '')::INTEGER,
     NULLIF(p.respondent_id, ''),
-    NULLIF(p.loan_type, '')::TINYINT,
-    NULLIF(p.property_type, '')::TINYINT,
-    NULLIF(p.loan_purpose, '')::TINYINT,
-    NULLIF(p.owner_occupancy, '')::TINYINT,
+    NULLIF(p.loan_type, '')::SMALLINT,
+    NULLIF(p.property_type, '')::SMALLINT,
+    NULLIF(p.loan_purpose, '')::SMALLINT,
+    NULLIF(p.owner_occupancy, '')::SMALLINT,
     NULLIF(p.loan_amount_000s, '')::NUMERIC,
-    NULLIF(p.preapproval, '')::TINYINT,
-    NULLIF(p.action_taken, '')::TINYINT,
+    NULLIF(p.preapproval, '')::SMALLINT,
+    NULLIF(p.action_taken, '')::SMALLINT,
     loc.location_id,
-    NULLIF(p.applicant_ethnicity, '')::TINYINT,
-    NULLIF(p.co_applicant_ethnicity, '')::TINYINT,
-    NULLIF(p.applicant_sex, '')::TINYINT,
-    NULLIF(p.co_applicant_sex, '')::TINYINT,
+    NULLIF(p.applicant_ethnicity, '')::SMALLINT,
+    NULLIF(p.co_applicant_ethnicity, '')::SMALLINT,
+    NULLIF(p.applicant_sex, '')::SMALLINT,
+    NULLIF(p.co_applicant_sex, '')::SMALLINT,
     NULLIF(p.applicant_income_000s, '')::NUMERIC,
-    NULLIF(p.purchaser_type, '')::TINYINT,
+    NULLIF(p.purchaser_type, '')::SMALLINT,
     NULLIF(p.rate_spread, ''),
-    NULLIF(p.hoepa_status, '')::TINYINT,
-    NULLIF(p.lien_status, '')::TINYINT,
-    NULLIF(p.edit_status, '')::TINYINT,
+    NULLIF(p.hoepa_status, '')::SMALLINT,
+    NULLIF(p.lien_status, '')::SMALLINT,
+    NULLIF(p.edit_status, '')::SMALLINT,
     NULLIF(p.sequence_number, ''),
-    NULLIF(p.application_date_indicator, '')::TINYINT
+    NULLIF(p.application_date_indicator, '')::SMALLINT
 FROM Preliminary p
 JOIN temp_location_id loc ON
     CONCAT(
